requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.0.7
django-cors-headers==4.4.0
django-phone-field==1.8.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
paypalrestsdk==1.13.3
phonenumberslite==8.13.42
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
python-magic==0.4.27
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
